+++ src/com/jpexs/decompiler/flash/gui/FolderListPanel.java
@@ -17,6 +17,7 @@
 package com.jpexs.decompiler.flash.gui;
 
 import com.jpexs.decompiler.flash.IdentifiersDeobfuscation;
+import com.jpexs.decompiler.flash.abc.ScriptPack;
 import com.jpexs.decompiler.flash.configuration.Configuration;
 import com.jpexs.decompiler.flash.gui.tagtree.AbstractTagTree;
 import com.jpexs.decompiler.flash.tags.DoInitActionTag;
@@ -251,6 +252,14 @@ public class FolderListPanel extends JPanel {
                     }
                     
                     Icon icon = ICONS.get(type);
+                    
+                    if (treeItem instanceof ScriptPack) {
+                        ScriptPack sp = (ScriptPack) treeItem;            
+                        if (sp.isDocumentClass()) {
+                            icon = View.getIcon("asclassmain32");                            
+                        }
+                    }
+                    
                     icon.paintIcon(l, g, x * CELL_WIDTH + BORDER_SIZE + PREVIEW_SIZE / 2 - icon.getIconWidth() / 2, y * CELL_HEIGHT + BORDER_SIZE + PREVIEW_SIZE / 2 - icon.getIconHeight() / 2);
                     String s = null;
                     if (treeItem instanceof DoInitActionTag) {